requirements.txt
# Dependencias requeridas
arabic-reshaper
asgiref
asn1crypto
certifi
cffi
charset-normalizer
cryptography
django_select2
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
html5lib==1.1
idna==3.10
inflection==0.5.1
lxml==6.0.0
mysqlclient==2.2.7
numpy==1.26.3  # Requerido para optimizaciones vectorizadas
numba>=0.58.0  # JIT compilation para optimización
django-redis>=5.4.0  # Cache y sesiones con Redis
openpyxl==3.0.10
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
PyMySQL==1.1.1
pypdf==5.9.0
python-bidi==0.6.6
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17

# Dependencias para testing
pytest==8.0.0
pytest-django>=4.5.0  # Testing para Django
hypothesis==6.99.4

# Dependencias opcionales para optimización
# Estas librerías mejoran el rendimiento pero no son estrictamente necesarias
joblib==1.4.2  # Paralelización mejorada

# Dependencias opcionales adicionales (no instaladas por defecto)
# Para instalarlas: pip install -r requirements-optional.txt
